+++ client/src/app/dashboard/resource/page.tsx
@@ -4,8 +4,8 @@ import { IoClose } from "react-icons/io5";
 import { FaPlus } from "react-icons/fa";
 import { useRouter } from "next/navigation";
 import { FiEdit, FiTrash2 } from "react-icons/fi";
-import { log } from "node:console";
-
+import Alerts from "../Alerts";
+import Loading from "../loading";
 const page = () => {
   const [uni_id, setUni_id] = useState("");
   const [resource_id, setResource_id] = useState(0);
@@ -18,6 +18,13 @@ const page = () => {
       duration: 0,
     },
   ]);
+  const [alertData, setAlertData] = useState({
+    status: 0,
+    function_name: "",
+    isModalOpen: false,
+    onConfirm: (confirm: boolean) => {},
+  });
+  const [loading, setLoading] = useState(true);
 
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [inputData, setInputData] = useState({
@@ -53,6 +60,7 @@ const page = () => {
     const data = await response.json();
 
     if (Array.isArray(data.data)) {
+      setLoading(false);
       setResource(data.data);
     } else {
       setResource([]);
@@ -60,25 +68,50 @@ const page = () => {
   };
 
   const handle_delete = async (res_id: any) => {
-    try {
-      const response = await fetch(
-        `http://localhost:3000/api/university/${uni_id}/resource/${res_id}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
+    setAlertData({
+      status: 1,
+      function_name: "delete",
+      isModalOpen: true,
+      onConfirm: async (confirm: boolean) => {
+        if (confirm) {
+          setLoading(true);
+          try {
+            const response = await fetch(
+              `http://localhost:3000/api/university/${uni_id}/resource/${res_id}`,
+              {
+                method: "DELETE",
+                headers: {
+                  "Content-Type": "application/json",
+                },
+              }
+            );
+            const data = await response.json();
+            if (data.status === 201) {
+              setLoading(false);
+              setAlertData({
+                status: 201,
+                function_name: "delete",
+                isModalOpen: true,
+                onConfirm: () => {},
+              });
+              setTimeout(() => {
+                setAlertData({
+                  status: 0,
+                  function_name: "",
+                  isModalOpen: false,
+                  onConfirm: () => {},
+                });
+              }, 3000);
+              console.log("Resource deleted successfully");
+              setResource(resource.filter((res) => res.id !== res_id));
+              router.refresh();
+            }
+          } catch (error) {
+            console.error(error);
+          }
         }
-      );
-      const data = await response.json();
-      if (data.status === 201) {
-        console.log("Resource deleted successfully");
-        setResource(resource.filter((res) => res.id !== res_id));
-        router.refresh();
-      }
-    } catch (error) {
-      console.error(error);
-    }
+      },
+    });
   };
 
   const handle_edit = (
@@ -118,28 +151,54 @@ const page = () => {
         }),
       });
       const data = await response.json();
-
       console.log(data);
-      
 
       if (data.function_name === "update_resource") {
+        setAlertData({
+          status: 201,
+          function_name: "update",
+          isModalOpen: true,
+          onConfirm: (confirm: boolean) => {},
+        });
+        setTimeout(() => {
+          setAlertData({
+            status: 0,
+            function_name: "",
+            isModalOpen: false,
+            onConfirm: () => {},
+          });
+        }, 3000);
         setResource((prev: any) =>
           prev.map((res: any) =>
             res.id === resource_id
               ? {
-                ...res,
-                resource_name: inputData.resource_name,
-                resource_type: inputData.resource_type,
-                capacity: inputData.capacity,
-                duration: inputData.duration,
-              }
+                  ...res,
+                  resource_name: inputData.resource_name,
+                  resource_type: inputData.resource_type,
+                  capacity: inputData.capacity,
+                  duration: inputData.duration,
+                }
               : res
           )
         );
       }
 
       if (data.function_name === "create_resource") {
-        setResource((prev) => [...prev, data.data[0]]);
+        setAlertData({
+          status: 201,
+          function_name: "create",
+          isModalOpen: true,
+          onConfirm: (confirm: boolean) => {},
+        });
+        setTimeout(() => {
+          setAlertData({
+            status: 0,
+            function_name: "",
+            isModalOpen: false,
+            onConfirm: () => {},
+          });
+        }, 3000);
+        setResource((prev) => [...prev, data.data]);
       }
 
       setIsModalOpen(false);
@@ -200,7 +259,7 @@ const page = () => {
 
   return (
     <div className="flex flex-col gap-6 justify-center items-center p-5 w-full">
-      <div className="flex justify-between w-full">
+      <div className="flex justify-between w-full h-full">
         <div></div>
         <div className="text-3xl font-bold text-slate-950">Resource</div>
         <button
@@ -211,7 +270,7 @@ const page = () => {
         </button>
 
         {isModalOpen && (
-          <div className="fixed z-10 top-0 left-0 right-0 bottom-0 bg-gray-900 bg-opacity-50 flex justify-center items-center">
+          <div className="fixed z-10 top-0 left-0 right-0 bottom-0 bg-gray-900 bg-opacity-50 flex justify-center items-center h-full">
             <div className="bg-white p-6 rounded-lg shadow-lg w-1/3 text-gray-950">
               <div className="flex relative">
                 <div className="flex-1 ">
@@ -287,12 +346,20 @@ const page = () => {
                 <div className="group flex relative mt-4">
                   <button
                     className="bg-blue-600 text-white px-4 py-2 rounded-md disabled:opacity-45 disabled:cursor-not-allowed"
-                    disabled={!inputData.resource_name || !inputData.resource_type || !inputData.capacity || !inputData.duration}
+                    disabled={
+                      !inputData.resource_name ||
+                      !inputData.resource_type ||
+                      !inputData.capacity ||
+                      !inputData.duration
+                    }
                   >
                     <span>Submit</span>
                   </button>
                   {/** Tooltip displayed only when the button is disabled and hovered */}
-                  {(!inputData.resource_name || !inputData.resource_type || !inputData.capacity || !inputData.duration) && (
+                  {(!inputData.resource_name ||
+                    !inputData.resource_type ||
+                    !inputData.capacity ||
+                    !inputData.duration) && (
                     <span
                       className="group-hover:opacity-100 transition-opacity bg-slate-500 px-1 
       text-sm text-gray-100 rounded-md absolute left-1/2 
@@ -307,9 +374,36 @@ const page = () => {
           </div>
         )}
       </div>
-      <div className="grid grid-cols-4 w-full gap-5">
-        {resource.length > 0 ? get_res_data : null}
-      </div>
+      {alertData.isModalOpen && (
+        <Alerts
+          status={alertData.status}
+          isModalOpen={alertData.isModalOpen}
+          function_name={alertData.function_name}
+          onConfirm={(confirm) => {
+            alertData.onConfirm(confirm);
+            setAlertData({
+              status: 0,
+              function_name: "",
+              isModalOpen: false,
+              onConfirm: () => {},
+            });
+          }}
+        ></Alerts>
+      )}
+
+      {loading ? (
+        <Loading />
+      ) : (
+        <div className="grid grid-cols-4 w-full gap-5">
+          {get_res_data.length > 0 ? (
+            get_res_data
+          ) : (
+            <div className="text-lg font-bold text-slate-950">
+              No Branch Found
+            </div>
+          )}
+        </div>
+      )}
     </div>
   );
 };
